requirements.txt
APScheduler==3.11.0
blinker==1.9.0
click==8.2.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
pytz==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
